requirements.txt
alembic==1.7.3
astroid==2.2.5
autopep8==1.5.7
Click==7.0
ecdsa==0.13.2
Flask==1.0.2
Flask-Cors==3.0.8
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.0
future==0.17.1
greenlet==1.1.0
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.0
Mako==1.1.5
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.9.1
pyasn1==0.4.8
pycodestyle==2.7.0
pycryptodome==3.3.1
pylint==2.3.1
python-dotenv==0.19.0
python-jose==3.3.0
python-jose-cryptodome==1.3.2
rsa==4.7.2
six==1.12.0
SQLAlchemy==1.4.22
toml==0.10.2
typed-ast==1.4.2
Werkzeug==0.15.5
wrapt==1.11.1
